DevFest25-WS/P5-RouterAgent/pyproject.toml
[project]
name = "p5-routeragent"
version = "0.1.0"
description = "Part 5: Router Agent - Intelligent Request Delegation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.19.0",
    "google-generativeai",
    "python-dotenv",
]

